backend/pyproject.toml
[project]
name = "mate-hackaton"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.68",
    "fal-client>=0.13.1",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "loguru>=0.7.3",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
